refactor: 미사용 메서드 제거 및 순서 컨벤션 적용
