+++ backend/src/test/java/com/woowacourse/gongcheck/acceptance/AcceptanceTest.java
@@ -8,7 +8,7 @@ import com.woowacourse.gongcheck.auth.application.response.TokenResponse;
 import com.woowacourse.gongcheck.auth.presentation.request.GuestEnterRequest;
 import com.woowacourse.gongcheck.core.application.ImageUploader;
 import com.woowacourse.gongcheck.core.application.NotificationService;
-import com.woowacourse.gongcheck.core.application.RunningTaskSseEmitterContainer;
+import com.woowacourse.gongcheck.core.domain.task.RunningTaskSseEmitterContainer;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
